rename other `data` folder
