+++ apps/meteor/ee/client/views/admin/engagementDashboard/dataView/LegendSymbol.stories.tsx
