+++ apps/meteor/ee/client/views/admin/engagementDashboard/dataView/usePeriodLabel.ts
@@ -1,6 +1,6 @@
+import { useTranslation } from '@rocket.chat/ui-contexts';
 import { useMemo } from 'react';
 
-import { useTranslation } from '@rocket.chat/ui-contexts'
 import { getPeriod, Period } from './periods';
 
 export const usePeriodLabel = (period: Period['key']): string => {
