+++ apps/meteor/ee/client/views/admin/engagementDashboard/users/ActiveUsersSection.tsx
@@ -8,8 +8,8 @@ import React, { ReactElement, useMemo } from 'react';
 import CounterSet from '../../../../../../client/components/dataView/CounterSet';
 import { useFormatDate } from '../../../../../../client/hooks/useFormatDate';
 import Section from '../Section';
-import DownloadDataButton from '../data/DownloadDataButton';
-import LegendSymbol from '../data/LegendSymbol';
+import DownloadDataButton from '../dataView/DownloadDataButton';
+import LegendSymbol from '../dataView/LegendSymbol';
 import { useActiveUsers } from './useActiveUsers';
 
 type ActiveUsersSectionProps = {
